Implemented EnumerateStuff methods on MockDirectoryInfo:  - EnumerateDirectories()  - EnumerateDirectories(string)  - EnumerateDirectories(string, SearchOption)  - EnumerateFiles()  - EnumerateFiles(string)  - EnumerateFiles(string, SearchOption)  - EnumerateFileSystemInfos()  - EnumerateFileSystemInfos(string)  - EnumerateFileSystemInfos(string, SearchOption)
Every one of these is simply a call to the equivalent GetStuff method and immediately returning the result with no extra work. I have *replaced* the implementation of Enumerate files so that it calls out to GetFiles, the previous implementation seemed to be broken - as far as I can tell it would always return *all* files!

Added test coverage for most of these methods, which essentially just consisted of copying and pasting the appropriate GetStuff method and modifying it to call EnumerateStuff instead. Changed the MockDirectoryInfo_EnumerateFiles_ShouldReturnAllFiles test method so that it has some additional files (i.e. ones which should *not* be returned).

Signed-off-by: Martin Evans <martindevans@gmail.com>