Leave NuGet version consistent
The mid-build upgrade was an attempt to work around an AppVeyor
publishing issue, however the AppVeyor NuGet provider doesn't use
the command line app by default anyway. The AppVeyor environment
has subseqently been upgraded to fix this properly, and we no longer
get 400 errors when pushing to the NuGet service.