TestingHelpers: Lazily create File|DirectorySecurity for better support on non-Windows OSes
Both FileSecurity and DirectorySecurity are not supported on non-Windows OSes and will throw PlatformNotSupportedException when called. MockDirectoryData & FileDirectoryData create instances of them when they are created. This makes them unusable on non-Windows OSes even if the actual *Security functionality is not used.

This change delays their creation to the point where they are actually necessary. That makes them usable again as long as *Security functioanlity is not used.