Add xml documentation inheritdoc conversion [#245] (#337)
* Generate documentation file for all builds

Setting the XML documentation file through Project Properties/Build
only generated the XML documentation file for NET40 and not
netstandard1.4 nor netstandard2.0. Adding the property
GenerateDocumentationFile and setting it to true generates the
documentation for all build types.

* Install inheritdoc

This is used to update the xml documentation by replacing the
inheritdoc tags with the associated documentation.

* Add xml conversion step to appveyor deployment

This should convert the inheritdoc tags to the documentation in the
mscorlib.xml file.

* Fix environment variable syntax

Fix the environment variable syntax for ps (powershell) rather than cmd.

* Fix xml documentation build step

Calling InheritDoc needed to occur before the nuget package creation,
so was moved to the before_build step. PowerShell variables are not
resolved if wrapped by single quotation marks, so these have been
removed.

* Shift InheritDoc call to before_package

The InheritDoc call is not working, hopefully this will fix it.

* Remove InheritDoc nuget package

The plan is to install InheritDoc temporarily using appveyor so that
it does not get deployed with System.IO.Abstractions.

* Add inheritdoc install to appveyor

* Add missing colon to install stage

* Add OutputDirectory to inheritdoc install

* Fix InheritDoc directory in before_package

Attempting to identify the folder InheritDoc is installed into.

* Update InheritDoc version

This is in attempt to fix the missing inheritdoc conversions.

* Add environment variables for inheritdoc install

* Fix environment variable syntax for nuget command

* Attempt to convert environment variables

The powershell call is not converting the environment variables to
their actual values. Adding the quotation marks might help.

* Change paths to use forward slashes

The backslashes were causing parsing errors.

* Attempt to parse environment variables

* Remove redundant quotation mark

* Remove quotation marks

* Adjust quotation marks

* Wrap InheritDoc version in quotation marks

* Remove quotation marks from path

* Revert to using backslashes

* Attempt to parse environment variables

Trying a non-powershell approach.
If this doesn't work, then back to the forums.

* Revert back to powershell variable type

* Amend InheritDoc call

This is based on feedback from the appveyor forum.