Add IFileSystem property to base classes
This allows people to get at the underlying IFileSystem implementation when developing extension methods.

For example, it may make sense to put your extension method on DirectoryBase, but you may require FileBase to actually implement it. This makes that possible.

In addition it fixes a bug in MockFileInfo.Replace where the returned FileInfo was for the concrete FileSystem not the mock one (bug caused by implicit casting).

I think this commit actually removes the need for supporting implicit casting at all.