Implemented MockFile.Create methods (#368)
Fixes #331
Fixes #332
Fixes #333

Implemented MockFile.Create methods as discussed in referenced issues.

No unit test for the overload that just takes a buffer size as it's ignored.

MockFileStream now receives a FileOptions which it checks on close to see if it should delete the file (FileOptions.DeleteOnClose) or encrypt it (FileOptions.Encrypted).

The overload that takes FileSecurity doesn't do anything with the argument, just passes it on. No unit test for this overload.