Bump base version to 3.0
With the new versioning strategy introduced in #369 our `AssemblyVersion` consists of only the 
major version. With the current base version of 2.2 this will be produce 2.0.0.0. This number is 
less than the currently released 2.1.0.256 which will pretty sure cause downstream problems.

This change bumps the base version to 3.0 so that we get 3.0.0.0.