Reduce target frameworks (#540)
This change reduces the target frameworks this library supports to .NET Standard 2.0 and 2.1.

The goal of this change is to reduce maintenance efforts and get rid of legacy workarounds.

As this is a breaking change for consumers (e.g. no more support from .NET Framework < 4.7) the major version is bumped to 8.

If you need support for older frameworks feel free to stay at v7.1.10 or fork this repo from 86066e0.