make MockFileData.NullObject internal
The `MockFileData.NullObject` was introduced in commit
0d85da38e12fd9baaeae178f739bc48d76afb4 as a way to represent the data
of a non-existing file. Nevertheless, the issue #830 shows that at least
one person used it as "some arbitrary MockFileData of an existing file
I don't care about".

BREAKING CHANGE: Make `MockFiledata.NullObject` an internal
implementation detail.

Reason: As we are going to really make this the data representing a
non-existing file by fixing/changing the `Attribute` property in one of
the next commits, make this property internal and thereby breaking to
force all consumers to really think about their tests and if they
really ment a non-existing file or simply some arbirary data.
Otherwise, the upcoming change would have nasty side effects if the
NullObject is used for arbitrary data of an existing file.

An alternative considered would be to rename this static property (which
is also breaking) to any users, but decided against it as I couldn't
think of a usecase where explicitly creating a file or directory which
doesn't exist instead of relying on what the library will for you when
being asked for a non-existing file.

E.g. it is the difference of

  var fs = new MockFileSystem();
  var nonExistingPath = XFS.Path(@"c:\some\non\existing\file.txt");
  fs.AddFile(nonExistingPath, MockFileData.NullObject);
  var fi = fs.FileInfo.FromFilename(nonExistingPath);

vs simply relying on the library to return the correct data for the
non-existing file:

  var fs = new MockFileSystem();
  var nonExistingPath = XFS.Path(@"c:\some\non\existing\file.txt");
  var fi = fs.FileInfo.FromFilename(nonExistingPath);

All tests are adapted to no longer use the MockFileData.NullObject.