fix: return -1 in MockFileInfo.Attributes and MockDirectoryInfo
According to the Microsoft documentation[1] the `Attributes` property of
FileSystemInfo (base of FileInfo and DirectoryInfo) returns
(FileAttributes)(-1) for a file/directory which doesn't exist when the
object instance is created. The on disk content is not looked at when
calling the `Attributes` property, only a pre-cached value is returned.
Only calling `Refresh` will read from disk.

To be on the safe side, I checked the documented behaviour with the real
implemenation with the below code snipped and it indeed shows the
documented behaviour.

    var fs = new FileSystem();
    var expected = (FileAttributes)(-1);
    var fi = fs.FileInfo.FromFileName(@"x:\not-existing\path");
    Assert.That(fi.Attributes, Is.EqualTo(expected));

This lead to quite some needed code adaptions, especially in the area of
the `GetMockFileDataForRead` methods, as those now no longer throw an
exception. Furthermore, as cachedMockFileData is always a non-null instance,
get rid of the logic checking for null in `GetMockFileDataForRead` and
`MockFileInfo.CopyTo(string destFileName)` method.

Also ensure that the used MockFileData is always `Clone`ed  (and therefore
even the MockFileData.NullObject) so that we operate on our own copy
and don't get any updates done in the FileSystem.

With this new semantics the `MockDiretory.Exists` method requires special
treatment, as we now must check for a valid MockFileData by checking
if Attributes != -1.

[1]: https://docs.microsoft.com/en-us/dotnet/api/system.io.filesysteminfo.attributes?view=net-6.0#remarks