fix: make invalid UNC path error message platform-specific (#1351)
This PR updates the `InvalidUncPath` error message to be platform specific, referencing the form `\\server\share` in Windows environments, and `//server/share` in other environments.

The method now checks the environment's platform (`RuntimeInformation.IsOSPlatform(OSPlatform.Windows)`) and conditionally returns an `IOException` with the correctly-formed UNC path in the message, using `System.Runtime.InteropServices` in the `CommonExceptions` class.