fix: file stream sharing (#1363)
This PR adds stateful handling of file streams opened with the `FileShare.None` option. If a file stream is attempted to be opened whilst another file stream of the same file path and `FileShare.None` is in use (that is to say, has been opened and not yet disposed), an `IOException` will be thrown and the creation of the second file stream will be disallowed.